feat: add 'when' parameter to all commands
In order to use the common `when` command parameter in an orb command, the orb command needs to have its own `when` parameter.

Here I'm adding `when` to each command so that `on_success`, `on_fail` or `always` can be plumbed through.

This is useful in circumstances where you may need to, say, install the Github CLI as part of the steps taken after a CI failure.

Example circleci.yml:
```yaml
commands:
    build-failure-pr-comment:
        steps:
            - gh/setup:
                when: on_fail
            - run: gh pr comment ...
                when: on_fail
```