Allow unauthenticated packages to be installed
This is a proposed work around to the expired GPG issue with the GitHub CLI packages - https://github.com/cli/cli/issues/9569.

The better fix is to support new keys, but this works and should be relatively safe as we know and mostly trust the source of the deb. 

It's a big improvement on the orb being broken as it is today.